# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HDHomeRun Signal API",
    "description": "Control and watch HDHomeRun tuners.\n\n## Event streams\n\n`GET /devices/{id}/tuner/{tuner}/stream` and `GET /devices/{id}/antenna/stream` are [Server-Sent Events](https://html.spec.whatwg.org/multipage/server-sent-events.html), which OpenAPI cannot describe as a stream, so they are documented here. Open one with `EventSource`; closing the connection ends the subscription, and `EventSource` reconnects by itself.\n\n* The tuner stream sends a `tuner-status` event about once a second, whose data is a `TunerEvent`.\n* The antenna stream sends an `antenna-mode-status` event about once a second, whose data is an array of `AntennaReading`, one per tuner.\n* The first reading is sent as soon as the stream opens, and a comment line (`: keepalive`) every 15 seconds keeps idle proxies from closing it.\n* A client may hold a limited number of streams (`HDHR_MAX_STREAMS_PER_CLIENT`, 16 by default); beyond that the answer is 429, and `EventSource` does not retry after one.\n\n## Versioning\n\nEverything is under `/api/v1`. The unversioned `/api/...` paths of earlier releases, and their Socket.IO endpoint, no longer exist.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/devices/{id}/antenna/stream": {
      "get": {
        "tags": [
          "streaming"
        ],
        "summary": "Watch every tuner of a device (event stream)",
        "description": "Server-Sent Events for aiming an antenna: an `antenna-mode-status` event about once a second, whose data is an array of `AntennaReading`, one per tuner in order.",
        "operationId": "antenna_stream",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "A device ID, IPv4 address or hostname.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tuners",
            "in": "query",
            "description": "How many tuners to watch, 1 to 8 (plain digits).",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The stream; each event is named `antenna-mode-status`.",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AntennaReading"
                  }
                }
              }
            }
          },
          "400": {
            "description": "A device or `tuners` is not valid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "This client already holds the most streams it may.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/devices/{id}/tuner/{tuner}/stream": {
      "get": {
        "tags": [
          "streaming"
        ],
        "summary": "Watch a tuner (event stream)",
        "description": "Server-Sent Events: a `tuner-status` event about once a second, whose data is a `TunerEvent`. See the notes on event streams in the API description.",
        "operationId": "tuner_stream",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "A device ID, IPv4 address or hostname.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tuner",
            "in": "path",
            "description": "The tuner index, 0 to 7.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32",
              "maximum": 7,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The stream; each event is named `tuner-status`.",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/TunerEvent"
                }
              }
            }
          },
          "400": {
            "description": "A device or tuner is not valid.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "This client already holds the most streams it may.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AntennaReading": {
        "type": "object",
        "description": "One tuner's reading inside an `antenna-mode-status` event.",
        "required": [
          "tuner"
        ],
        "properties": {
          "status": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TunerStatus",
                "description": "Null when that tuner did not answer."
              }
            ]
          },
          "tuner": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "The body of every error response.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "What went wrong, in words."
          }
        }
      },
      "PlpInfo": {
        "type": "object",
        "description": "One ATSC 3.0 physical layer pipe, from `/tuner<n>/plpinfo`.",
        "properties": {
          "coderate": {
            "type": [
              "string",
              "null"
            ]
          },
          "layer": {
            "type": [
              "string",
              "null"
            ]
          },
          "lls": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "lock": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "modulation": {
            "type": [
              "string",
              "null"
            ]
          },
          "sfi": {
            "type": [
              "string",
              "null"
            ]
          },
          "timeInterleaving": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "TunerEvent": {
        "type": "object",
        "description": "The payload of a `tuner-status` event: the tuner's status, plus the current\nprogram and the ATSC 3.0 details.\n\nA tuner that does not answer still produces an event, with none of the status\nfields (no `channel`, no `lock`), so a page can tell the device is not\nreporting. `plpInfo` and `l1Info` are only fetched while a channel is tuned.",
        "required": [
          "currentProgram",
          "plpInfo",
          "l1Info"
        ],
        "properties": {
          "bps": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Stream bitrate in bits per second.",
            "minimum": 0
          },
          "channel": {
            "type": [
              "string",
              "null"
            ],
            "description": "The tuned channel as the device reports it, e.g. `auto:34`, or `none`."
          },
          "currentProgram": {
            "type": [
              "string",
              "null"
            ],
            "description": "The selected program number, as the device reports it, or null."
          },
          "debugRaw": {
            "type": [
              "string",
              "null"
            ],
            "description": "The debug counters the dB estimates came from."
          },
          "l1Info": {
            "type": [
              "object",
              "null"
            ],
            "description": "ATSC 3.0 L1 signalling as raw key/value pairs, or null.",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "lock": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "Whether the tuner has a lock on the signal."
          },
          "plpInfo": {
            "type": [
              "object",
              "null"
            ],
            "description": "ATSC 3.0 PLPs by id, or null when the device reports none.",
            "additionalProperties": {
              "$ref": "#/components/schemas/PlpInfo"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "pps": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Packets per second.",
            "minimum": 0
          },
          "seq": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Symbol error quality, percent.",
            "minimum": 0
          },
          "snq": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Signal-to-noise quality, percent.",
            "minimum": 0
          },
          "snrDb": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Signal-to-noise estimated in dB, when the device reports the counters for it."
          },
          "ss": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Signal strength, percent.",
            "minimum": 0
          },
          "ssDb": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Signal strength estimated in dBm, when the device reports the counters for it."
          }
        }
      },
      "TunerStatus": {
        "type": "object",
        "description": "One reading of `/tuner<n>/status`.",
        "required": [
          "lock"
        ],
        "properties": {
          "bps": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Stream bitrate in bits per second.",
            "minimum": 0
          },
          "channel": {
            "type": [
              "string",
              "null"
            ],
            "description": "The tuned channel as the device reports it, e.g. `8vsb:27`, or `none`."
          },
          "debugRaw": {
            "type": [
              "string",
              "null"
            ],
            "description": "The debug counters the dB estimates came from, as `signal-snr/third`."
          },
          "lock": {
            "type": "boolean"
          },
          "pps": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Packets per second.",
            "minimum": 0
          },
          "seq": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Symbol error quality, percent.",
            "minimum": 0
          },
          "snq": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Signal-to-noise quality, percent.",
            "minimum": 0
          },
          "snrDb": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Signal-to-noise estimated in dB from the debug counters."
          },
          "ss": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Signal strength, percent.",
            "minimum": 0
          },
          "ssDb": {
            "type": [
              "number",
              "null"
            ],
            "format": "double",
            "description": "Signal strength estimated in dBm from the debug counters."
          }
        }
      },
...
    }
  },
...
}